// private/carrier/assets/gel-workspace/apps/company/src/api/corp/info/common.ts
export type CorpOrganizationType =
  // 香港
  | 'HK'
  // 个体工商户
  | 'IIP'
  // 公司
  | 'CO'
  // 机关
  | 'GOV'
  // 事业单位
  | 'SOE'
  // 社会团体
  | 'NGO'
  // 合伙企业
  | 'PE'
  // 农民专业合作社法人
  | 'FPC'
  // 个人独资企业
  | 'SPE'
  // 其他企业 内资非法人企业、内资非公司企业分支机构、内资分公司、外商投资企业分支机构、合伙企业分支机构、个人独资企业分支机构、外国<地区>企业在中国境内从事生产经营活动
  | 'OE'
  // 农民专业合作社分支机构
  | 'FCP'
  // 默认
  | '00'
  | string